// projects/msct-webui/src/app/services/types.ts
export enum MessageTypes {
    Cmd = 'cmd',
    Status = 'status',
    Testresult = 'testresult',
    Testresults = 'testresults',
    Usersettings = 'usersettings',
    Logdata = 'logs',
    Logfile = 'logfile',
    ConnectionId = 'connectionid',
    Yield = 'yield',
    Lotdata = 'lotdata',
    BinTable = 'bintable'
}